requirements.txt
Django==2.0.6
djangorestframework==3.8.2
python-dateutil==2.7.3
python-mimeparse==1.6.0
pytz==2018.5
six==1.11.0
